Remove failing assert
As of setuptools 77.0.3, this assert is no longer passing. This means
projects that require this for building cannot build, and projects that
use it at runtime (e.g. cothread's catools) cannot run.

It should be noted that this passes under setuptools 77.0.1, so there
is apparently a breaking change in a minor release.